chore(filter): move FilterInputDocs component
