+++ frontend/src/components/project/views/ViewEditForm.vue
@@ -9,7 +9,7 @@ import {useLabelStore} from '@/stores/labels'
 import {useProjectStore} from '@/stores/projects'
 
 import XButton from '@/components/input/Button.vue'
-import FilterInputDocs from '@/components/project/partials/FilterInputDocs.vue'
+import FilterInputDocs from '@/components/input/filter/FilterInputDocs.vue'
 import FilterInput from '@/components/input/filter/FilterInput.vue'
 
 const props = withDefaults(defineProps<{
